+++ lib/felix/config.rb
@@ -1,6 +1,8 @@
 ## Helper class that handles configs
 require "felix/metadata"
 
+## Processes loaded options. Doesn't actually store them, though.
+
 module PetitFelix
 
 	class	Config
@@ -14,7 +16,7 @@ module PetitFelix
 			"image_dir" => "./assets/images",
 			"input_files" => "./md/*",
 			"output_dir" => "./output",
-			"task" => "basic_pdf",
+			"task" => "pdf-single",
 		}
 
 		## Hash for custom command line argument calls
@@ -41,7 +43,7 @@ module PetitFelix
 
 		### Command Line Arguments
 
-		def self.cl_add_config(command,args, index, cl_args)
+		def self.cl_add_config command,args, index, cl_args
 			cl_args[command] = args[index + 1]
 		end
 
@@ -53,7 +55,7 @@ module PetitFelix
 		## then command line arguments,
 		## then finally any arguments defined in the metadata.
 
-		def load_config(wm, passed_args, args)
+		def load_config wm, passed_args, args
 		
 			@processed_arguments = []
 	
@@ -67,33 +69,37 @@ module PetitFelix
 			
 			default_config = {}
 			
-			if File.file?("./default.cfg")
+			if File.file? "./default.cfg" 
 				
-				default_config = metadata.get_metadata(File.read("./default.cfg"))
+				default_config = metadata.get_metadata(File.read "./default.cfg")
 			end
 			
 			# Loads command line arguments
-			cl_args = load_cl_args(args)
+			cl_args = load_cl_args args
 			
 			# Loads default worker options
 			worker = ""
 			worker_options = nil
 			
-			if default_options.key?("task")
+			# Gets the task to run.
+			
+			if default_options.key? "task"
 				worker = default_options["task"]
 			end
 			
-			if default_config.key?("task")
+			if default_config.key? "task"
 				worker = default_config["task"]
 			end
 			
-			if cl_args.key?("task")
+			if cl_args.key? "task"
 				worker = cl_args["task"]
 			end
 			
-			if passed_args.key?("task")
+			if passed_args.key? "task"
 				worker = passed_args["task"]
 			end
+			
+			# Loads the worker based on the task.
 			
 			if worker != ""
 				worker_options = Marshal.load(Marshal.dump(wm.get_task_options worker))
@@ -129,7 +135,7 @@ module PetitFelix
 		end
 
 		# Loads command line arguments
-		def load_cl_args(args)
+		def load_cl_args args
 			cl_args = {}
 			
 			index = 0
@@ -138,14 +144,14 @@ module PetitFelix
 			
 				command = com.downcase
 			
-				if command.start_with?("--")
+				if command.start_with? "--"
 					command = command[2..]
 				
-					if !@processed_arguments.include?(command)
+					if !@processed_arguments.include? command
 				
-						if CL_DATA.key?(command)
+						if CL_DATA.key? command
 						
-							CL_DATA[command].call(command, args, index, cl_args)
+							CL_DATA[command].call command, args, index, cl_args
 
 						end