+++ lib/felix/error.rb
@@ -1,8 +1,10 @@
+## Prints an error message.
+
 module PetitFelix
 	class Error
 	
-		def print_err(error)
-			print "\n\nError: " + error + "\n"
+		def print_err error
+			print "\n\nError: #{error}\n"
 		end
 		
 	end